Updated regex;port code
